ensure char is unsigned when encoding in percent

+++ url.cpp
@@ -462,7 +462,7 @@ class encode {
             if (is_char(c,e.m_mask))
                 o<<c;
             else
-                o<<'%'<<"0123456789ABCDEF"[c>>4]<<"0123456789ABCDEF"[c&0xF];
+                o<<'%'<<"0123456789ABCDEF"[((uint8_t)c)>>4]<<"0123456789ABCDEF"[((uint8_t)c)&0xF];
         return o;
     }
 };
@@ -489,7 +489,7 @@ class encode_query_key {
             else if (is_char(c,e.m_mask))
                 o<<c;
             else
-                o<<'%'<<"0123456789ABCDEF"[c>>4]<<"0123456789ABCDEF"[c&0xF];
+                o<<'%'<<"0123456789ABCDEF"[((uint8_t)c)>>4]<<"0123456789ABCDEF"[((uint8_t)c)&0xF];
         return o;
     }
 };
